fixup! Provide Arduino lib dependencies

+++ rpi/lib/arduino/Print.cpp
@@ -21,6 +21,7 @@
  Modified May 2015 by Michael C. Miller - ESP31B progmem support
  */
 
+#include <cstddef>
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -58,7 +59,7 @@ size_t Print::printf(const char *format, ...)
         va_end(arg);
         return 0;
     };
-    if(len >= sizeof(loc_buf)){
+    if((size_t)len >= sizeof(loc_buf)){
         temp = (char*) malloc(len+1);
         if(temp == NULL) {
             va_end(arg);
